+++ src/components/header/style.less
@@ -8,7 +8,7 @@
 	height: 56px;
 	padding: 0;
 	background: #006249;
-	box-shadow: 0 0 5px rgba(0, 0, 0, 0.5);
+	box-shadow: 0 3px 5px rgba(0, 0, 0, 0.5);
 	z-index: 50;
 
 	h1 {
