+++ src/components/main/style.less
